Update ykval-synclib.php
According to https://secure.php.net/manual/en/function.curl-multi-info-read.php, "The data the returned resource points to will not survive calling curl_multi_remove_handle()." Therefore, it is better to call curl_multi_remove_handle() after we're finished with the handle and its related info data returned from curl_multi_info_read(). Not doing so will make the info data unreadable. This could lead to logging when re-syncing that makes it appear as though syncing is not successful when in reality it is working just fine.